requirements.txt
# Core ML
torch>=2.1.0
onnxruntime>=1.16.0
onnx>=1.15.0
einops>=0.7.0

# Data
polars>=0.20.0
numpy>=1.24.0
duckdb>=0.9.0
pyarrow>=14.0.0
pandas-ta>=0.3.14b

# NLP / Sentiment
transformers>=4.35.0
tokenizers>=0.15.0

# RL
stable-baselines3>=2.2.0
gymnasium>=0.29.0

# Config & Tracking
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.16.0
python-dotenv>=1.0.0

# Monitoring
prometheus-client>=0.19.0
python-telegram-bot>=20.7

# Web scraping (FT.com)
curl_cffi>=0.7.0
playwright>=1.40.0

# Utilities
requests>=2.31.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tqdm>=4.66.0
loguru>=0.7.0
joblib>=1.3.0
